Custom command and event handler added
I added my complete custom command and event handler, adding all
previous external files into the ready event. All commands have been
changed to work with the command handler but you may want to check that
there is nothing missing from the imports, just require anything that is
missing above the module.exports.run line, also I added a few lines of
code which use the 'chalk' dependecy so add that if you want but it
isn't needed, you will just need to change some lines if you don't add
it.
