+++ This I changed.md
@@ -1,0 +1,2 @@
+Added the chalk dependency
+Moved entire `setPresence.js` into the ready event
