--- events/guildEvents/guildCreate.js
@@ -1,5 +1,0 @@
-function guildCreate(guild, rethinkdb){
-    
-}
-
-module.exports = guildCreate;
