--- events/messageEvent/addXP.js
@@ -1,16 +1,0 @@
-function addXP(msg, discord, client, rethinkdb, config){
-
-    /*var info = {};
-    var info1 = {};
-    info1['level'] = "2";
-    info[msg.author.id] = info1;
-
-    rethinkdb.db('horizon').table('levels').insert(
-      //CODE GOES HERE
-      info
-    ).run(global.connection, (err, info) => {
-      if (err) {
-        console.log(err);
-    }});*/
-}
-module.exports = addXP;
