--- events/messageEvent/commands/musicCommands/pauseCommand.js
@@ -1,54 +1,0 @@
-function pauseCommand(msg, discord, client, config, servers, ytdl){
-    if(msg.member.voiceChannel && msg.guild.voiceConnection && msg.member.voiceChannelID == msg.guild.voiceConnection.channel.id){
-        var dispatcher = msg.member.voiceChannel.connection.dispatcher;
-        var embed;
-        var info = ytdl.getInfo(servers[msg.guild.id].queue[0], (err, info) =>{
-            if (err) throw err;
-            if(msg.member.hasPermission('ADMINISTRATOR')){
-                embed = new discord.RichEmbed()
-                //TITLE
-                .setAuthor("Music")
-                .setColor('#FF6800')
-    
-                //INFO
-                .addField(':pause_button: Song Paused', '**' + msg.author.username + '** has paused **' + info.title + '**')
-        
-    
-                //FOOTER
-                .setFooter("Sent by " + client.user.username.toString())
-                .setTimestamp();
-                msg.channel.send({embed});
-                dispatcher.pause();
-            }else{
-                embed = new discord.RichEmbed()
-                //TITLE
-                .setAuthor("Music")
-                .setColor('#FF6800')
-        
-                //INFO
-                .addField(':x: Pause Error', 'You do not have permission to pause the music. \nPermission Needed: ADMINISTRATOR')
-                
-        
-                //FOOTER
-                .setFooter("Sent by " + client.user.username.toString())
-                .setTimestamp();
-                msg.channel.send({embed});
-            }
-        });
-    }else{
-        embed = new discord.RichEmbed()
-        //TITLE
-        .setAuthor("Music")
-        .setColor('#FF6800')
-
-        //INFO
-        .addField(':x: Pause Error', 'Either you or the bot is not in a voice channel or \nyoure not in the same voice channel as the bot.')
-        
-
-        //FOOTER
-        .setFooter("Sent by " + client.user.username.toString())
-        .setTimestamp();
-        msg.channel.send({embed});
-    }
-}
-module.exports = pauseCommand;
