+++ storage/events/guildCreate.js
@@ -1,0 +1,5 @@
+function guildCreate(guild, rethinkdb){
+    
+}
+
+module.exports = guildCreate;
