+++ storage/util/eventLoader.js
@@ -1,0 +1,6 @@
+const reqEvent = (event) => require(`../events/${event}`);
+module.exports = client => {
+  client.on('ready', () => reqEvent('ready.js')(client));
+  client.on('message', reqEvent('message.js'));
+  client.on('guildCreate', reqEvent('guildCreate.js'));
+};
